requirements.txt
flask==2.3.3
gunicorn==21.2.0
werkzeug==2.3.7
mcp-server>=0.2.0
PyYAML==6.0.1
requests==2.31.0
python-dotenv>=1.0.0
uvicorn>=0.22.0
starlette>=0.36.3
httpx>=0.24.0
mcp>=0.1.2
fastapi>=0.100.0
pydantic>=2.0.0 